@@ -13,17 +13,18 @@
   "private": true,
   "dependencies": {
     "@amcharts/amcharts4": "^4.9.13",
-    "@angular/animations": "^8.1.1",
-    "@angular/common": "^8.1.1",
-    "@angular/compiler": "^8.1.1",
-    "@angular/core": "^8.1.1",
-    "@angular/fire": "^5.4.2",
-    "@angular/forms": "^8.1.1",
-    "@angular/platform-browser": "^8.1.1",
-    "@angular/platform-browser-dynamic": "^8.1.1",
-    "@angular/platform-server": "^8.1.1",
-    "@angular/router": "^8.1.1",
+    "@angular/animations": "^15.2.10",
+    "@angular/common": "^15.2.10",
+    "@angular/compiler": "^15.2.10",
+    "@angular/core": "^15.2.10",
+    "@angular/fire": "^17.0.0",
+    "@angular/forms": "^15.2.10",
+    "@angular/platform-browser": "^15.2.10",
+    "@angular/platform-browser-dynamic": "^15.2.10",
+    "@angular/platform-server": "^15.2.10",
+    "@angular/router": "^15.2.10",
     "@fortawesome/fontawesome-free": "^5.5.0",
+    "@techiediaries/ngx-qrcode": "^9.1.0",
     "@types/chartist": "^0.9.43",
     "ag-grid-angular": "^21.1.0",
     "ag-grid-community": "^21.1.1",
@@ -33,60 +34,60 @@
     "chartist": "^0.11.0",
     "core-js": "^2.5.7",
     "date-fns": "^2.28.0",
-    "express": "^4.18.2",
+    "express": "^4.18.3",
     "express-basic-auth": "^1.2.1",
-    "firebase": "^7.24.0",
-    "firebase-functions": "^4.2.1",
-    "firebase-tools": "^10.2.0",
+    "firebase": "^10.2.0",
+    "firebase-functions": "^4.7.0",
     "html2canvas": "^1.4.1",
     "jquery": "^3.3.1",
     "jspdf": "^2.5.1",
-    "jspdf-autotable": "^3.5.30",
+    "jspdf-autotable": "^3.8.2",
+    "ng": "^0.0.0",
     "ng-chartist": "^1.1.1",
     "ng-chartjs": "0.0.7",
-    "ng-zorro-antd": "^8.5.2",
+    "ng-zorro-antd": "^15.0.3",
     "ngx-csv-parser": "0.0.3",
     "ngx-highlightjs": "^2.1.1",
     "ngx-infinite-scroll": "^6.0.1",
     "ngx-perfect-scrollbar": "^7.2.1",
     "ngx-permissions": "^7.0.3",
-    "ngx-qrcode2": "0.0.9",
     "ngx-quill": "^4.6.3",
     "ngx-sortablejs": "^3.1.4",
     "quill": "^1.3.6",
     "rxjs": "^6.3.3",
     "rxjs-compat": "^6.3.3",
     "sortablejs": "^1.10.2",
+    "update": "^0.7.4",
     "webpack-bundle-analyzer": "^3.0.3",
-    "zone.js": "~0.9.1"
+    "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/architect": "<0.900 || ^0.900.0-0 || ^9.0.0-0",
-    "@angular-devkit/build-angular": "~0.801.1",
-    "@angular/cli": "^8.1.1",
-    "@angular/compiler-cli": "^8.1.1",
-    "@angular/language-service": "^8.1.1",
+    "@angular-devkit/architect": "0.1502.10",
+    "@angular-devkit/build-angular": "^15.2.10",
+    "@angular/cli": "^15.2.10",
+    "@angular/compiler-cli": "^15.2.10",
+    "@angular/language-service": "^15.2.10",
     "@types/jasmine": "^3.0.0",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "^10.12.9",
-    "@types/sortablejs": "^1.10.4",
+    "@types/sortablejs": "^1.15.8",
     "codelyzer": "^5.0.1",
-    "firebase-tools": "^7.12.0",
+    "firebase-tools": "^12.9.1",
     "fuzzy": "^0.1.3",
     "inquirer": "^6.2.2",
     "inquirer-autocomplete-prompt": "^1.0.1",
-    "jasmine-core": "^3.3.0",
-    "jasmine-spec-reporter": "^4.2.1",
-    "karma": "^3.1.1",
-    "karma-chrome-launcher": "^2.2.0",
+    "jasmine-core": "~3.5.0",
+    "jasmine-spec-reporter": "~5.0.0",
+    "karma": "~6.4.3",
+    "karma-chrome-launcher": "~3.1.0",
     "karma-cli": "^1.0.1",
-    "karma-coverage-istanbul-reporter": "~2.0.0",
-    "karma-jasmine": "^2.0.1",
-    "karma-jasmine-html-reporter": "^1.4.0",
-    "protractor": "^5.4.1",
+    "karma-coverage-istanbul-reporter": "~3.0.2",
+    "karma-jasmine": "~4.0.0",
+    "karma-jasmine-html-reporter": "^1.5.0",
+    "protractor": "~7.0.0",
     "rxjs-tslint": "^0.1.5",
     "ts-node": "~5.0.1",
-    "tslint": "^5.11.0",
-    "typescript": "3.4.5"
+    "tslint": "~6.1.0",
+    "typescript": "4.9.5"
   }
 }
