@@ -6,7 +6,7 @@ describe('CustomersService', () => {
   beforeEach(() => TestBed.configureTestingModule({}));
 
   it('should be created', () => {
-    const service: CustomersService = TestBed.get(CustomersService);
+    const service: CustomersService = TestBed.inject(CustomersService);
     expect(service).toBeTruthy();
   });
 });
