@@ -1,10 +1,10 @@
 import { Routes } from '@angular/router';
-import { canActivate, redirectLoggedInTo } from '@angular/fire/auth-guard';
+import { AuthPipe, redirectLoggedInTo } from '@angular/fire/auth-guard';
 
 export const FullLayout_ROUTES: Routes = [
     {
         path: 'authentication',
         loadChildren: () => import('../../authentication/authentication.module').then(m => m.AuthenticationModule),
-        ...canActivate(redirectLoggedInTo(['/dashboard']))
+        canActivate: [redirectLoggedInTo(['/dashboard'])] as AuthPipe[],
     }
 ];
