@@ -1,12 +1,13 @@
 import { Injectable } from '@angular/core';
-import { AngularFirestore } from '@angular/fire/firestore';
-import * as firebase from 'firebase';
+import { AngularFirestore } from '@angular/fire/compat/firestore';
+import { increment, serverTimestamp } from 'firebase/firestore';
+
 @Injectable({
   providedIn: 'root'
 })
 export class AccountsService {
 
-  db = firebase.firestore();
+  //db = firebase.firestore();
 
   constructor(private afs: AngularFirestore) { }
 
@@ -29,40 +30,45 @@ export class AccountsService {
   }
 
   setAccount(account: any) {
-    const docId = this.afs.createId();
-    const increment = firebase.firestore.FieldValue.increment(1);
-    const last_updated = new firebase.firestore.Timestamp(new Date().getTime() / 1000, 0);
-    const newAccountRef = this.db.collection('customers').doc(docId);
-    const newPublicAccountRef = this.db.collection('pCustomers').doc(docId);
-    const stats = this.db.collection('summarized').doc('ZcVXcD1p4O7mYjwBIiLv');
-    const batch = this.db.batch();
-    batch.set(newAccountRef, account);
-    batch.set(newPublicAccountRef, { name: account.name, active: false });
-    batch.set(stats, { currentAccounts: increment, last_updated }, {merge: true});
+    const docId = this.afs.createId();  
+    const last_updated = serverTimestamp();
+    const newAccountRef = this.afs.collection('customers').doc(docId);
+    const newPublicAccountRef = this.afs.collection('pCustomers').doc(docId);
+    const stats = this.afs.collection('summarized').doc('ZcVXcD1p4O7mYjwBIiLv');
+    const batch = this.afs.firestore.batch();
+    
+    batch.set(newAccountRef.ref, account);
+    batch.set(newPublicAccountRef.ref, { name: account.name, active: false });
+    batch.set(stats.ref, { currentAccounts: increment(1), last_updated }, {merge: true});
+
     return batch.commit();
   }
 
   deleteAccount(accountId: string ) {
     const docId = accountId;
-    const decrement = firebase.firestore.FieldValue.increment(-1);
-    const last_updated = new firebase.firestore.Timestamp(new Date().getTime() / 1000, 0);
-    const accountRef = this.db.collection('customers').doc(docId);
-    const publicAccountRef = this.db.collection('pCustomers').doc(docId);
-    const stats = this.db.collection('summarized').doc('ZcVXcD1p4O7mYjwBIiLv');
-    const batch = this.db.batch();
-    batch.delete(accountRef);
-    batch.delete(publicAccountRef);
-    batch.set(stats, { currentAccounts: decrement, last_updated }, {merge: true});
+   
+    const last_updated = serverTimestamp();
+    const accountRef = this.afs.collection('customers').doc(docId);
+    const publicAccountRef = this.afs.collection('pCustomers').doc(docId);
+    const stats = this.afs.collection('summarized').doc('ZcVXcD1p4O7mYjwBIiLv');
+    const batch = this.afs.firestore.batch(); 
+    batch.delete(accountRef.ref);
+    batch.delete(publicAccountRef.ref);
+    batch.set(stats.ref, { currentAccounts: decrement(1), last_updated }, {merge: true});
     return batch.commit();
   }
 
   toggleAccountActive(accountId: string, active: boolean) {
     const docId = accountId;
-    const accountRef = this.db.collection('customers').doc(docId);
-    const publicAccountRef = this.db.collection('pCustomers').doc(docId);
-    const batch = this.db.batch();
-    batch.set(accountRef, { active: !active}, { merge: true});
-    batch.set(publicAccountRef, { active: !active}, { merge: true});
+    const accountRef = this.afs.collection('customers').doc(docId);
+    const publicAccountRef = this.afs.collection('pCustomers').doc(docId);
+    const batch = this.afs.firestore.batch();
+    batch.set(accountRef.ref, { active: !active}, { merge: true});
+    batch.set(publicAccountRef.ref, { active: !active}, { merge: true});
     return batch.commit();
   }
 }
+function decrement(arg0: number) {
+  throw new Error('Function not implemented.');
+}
+
