@@ -6,7 +6,7 @@ describe('DashboardService', () => {
   beforeEach(() => TestBed.configureTestingModule({}));
 
   it('should be created', () => {
-    const service: DashboardService = TestBed.get(DashboardService);
+    const service: DashboardService = TestBed.inject(DashboardService);
     expect(service).toBeTruthy();
   });
 });
