@@ -6,7 +6,7 @@ describe('FormatterService', () => {
   beforeEach(() => TestBed.configureTestingModule({}));
 
   it('should be created', () => {
-    const service: FormatterService = TestBed.get(FormatterService);
+    const service: FormatterService = TestBed.inject(FormatterService);
     expect(service).toBeTruthy();
   });
 });
